+++ .vscodeignore
@@ -4,12 +4,21 @@
 .*
 .autobuild/**
 src/**
+# Workspace packages ship as compiled output only; src is resolved via the
+# "#sl-ide-ws-client" subpath import in package.json, not through node_modules
+packages/*/src/**
+packages/*/node_modules/**
+packages/*/doc/**
 out/test/**
 data/dev/**
 doc/dev/**
 out_test/**
 # IMPORTANT: Do not exclude entire node_modules; needed runtime deps (js-yaml, ws, @iarna/toml)
 # Instead prune obvious dev-only artifacts selectively (keep dependency trees intact)
+# npm links workspaces here for local development; the packaged extension
+# resolves them through the "#sl-ide-ws-client" subpath import instead
+node_modules/@secondlife/sl-ide-ws-client
+node_modules/@secondlife/**
 node_modules/@types/**
 node_modules/mocha/**
 node_modules/@vscode/test-*/**
