+++ eslint.config.mjs
@@ -5,6 +5,9 @@ import tsPlugin from '@typescript-eslint/eslint-plugin';
 export default [
   js.configs.recommended,
   {
+    ignores: ['packages/*/dist/**'],
+  },
+  {
     files: ['src/**/*.ts'],
     ignores: ['src/webview/**/*.ts'],
     languageOptions: {
@@ -67,6 +70,62 @@ export default [
       // Allow console.log in development
       'no-console': 'off',
       // Common development patterns
+      'no-undef': 'warn',
+      'no-useless-catch': 'warn',
+      'no-unreachable': 'warn',
+      'no-case-declarations': 'warn',
+    },
+  },
+  {
+    files: ['packages/sl-ide-ws-client/src/**/*.ts'],
+    languageOptions: {
+      parser: tsParser,
+      parserOptions: {
+        ecmaVersion: 2022,
+        sourceType: 'module',
+        project: './packages/sl-ide-ws-client/tsconfig.json',
+      },
+      globals: {
+        console: 'readonly',
+        process: 'readonly',
+        Buffer: 'readonly',
+        setTimeout: 'readonly',
+        clearTimeout: 'readonly',
+        setInterval: 'readonly',
+        clearInterval: 'readonly',
+        require: 'readonly',
+        module: 'readonly',
+        exports: 'readonly',
+        NodeJS: 'readonly',
+      },
+    },
+    plugins: {
+      '@typescript-eslint': tsPlugin,
+    },
+    rules: {
+      ...tsPlugin.configs.recommended.rules,
+      '@typescript-eslint/no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
+      '@typescript-eslint/no-explicit-any': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'warn',
+      '@typescript-eslint/explicit-module-boundary-types': 'off',
+      '@typescript-eslint/no-non-null-assertion': 'off',
+      '@typescript-eslint/no-empty-object-type': 'off',
+      '@typescript-eslint/no-unused-expressions': 'warn',
+      'indent': ['error', 4, {
+        'SwitchCase': 1,
+        'VariableDeclarator': 1,
+        'outerIIFEBody': 1,
+        'MemberExpression': 1,
+        'FunctionDeclaration': { 'parameters': 1, 'body': 1 },
+        'FunctionExpression': { 'parameters': 1, 'body': 1 },
+        'CallExpression': { 'arguments': 1 },
+        'ArrayExpression': 1,
+        'ObjectExpression': 1,
+        'ImportDeclaration': 1,
+        'flatTernaryExpressions': false,
+        'ignoreComments': false
+      }],
+      'no-console': 'off',
       'no-undef': 'warn',
       'no-useless-catch': 'warn',
       'no-unreachable': 'warn',
