+++ packages/sl-ide-ws-client/src/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @file index.ts
+ * Public index for the websocket client.
+ */
+export * from "./events";
+export * from "./websockclient";
+export * from "./jsonrpcclient";
+export * from "./transport";
+export * from "./protocol";
+export * from "./objectcontentinterfaces";
+export * from "./objectcontentservice";
+export * from "./viewereditwsclient";
